lab1/requirements.txt
robotframework
robotframework-mongodblibrary=0.3.1
pymongo
requests